Board API => BoardDetail API로 일부 이동 및 조회수 저장 로직 리팩토링
