+++ src/main/java/com/bbangle/bbangle/board/controller/BoardController.java
@@ -1,34 +1,23 @@
 package com.bbangle.bbangle.board.controller;
 
-import com.bbangle.bbangle.board.dto.BoardImageDetailResponse;
 import com.bbangle.bbangle.board.dto.BoardResponseDto;
 import com.bbangle.bbangle.board.dto.FilterRequest;
-import com.bbangle.bbangle.board.dto.orders.ProductResponse;
 import com.bbangle.bbangle.board.service.BoardService;
-import com.bbangle.bbangle.board.service.ProductService;
 import com.bbangle.bbangle.common.dto.CommonResult;
 import com.bbangle.bbangle.common.service.ResponseService;
 import com.bbangle.bbangle.board.sort.FolderBoardSortType;
 import com.bbangle.bbangle.board.sort.SortType;
 import com.bbangle.bbangle.page.BoardCustomPage;
 import com.bbangle.bbangle.page.CustomPage;
-import com.bbangle.bbangle.review.dto.SummarizedReviewResponse;
-import com.bbangle.bbangle.review.service.ReviewService;
-import com.bbangle.bbangle.store.dto.StoreDto;
-import com.bbangle.bbangle.store.service.StoreService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Content;
 import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.tags.Tag;
-import jakarta.servlet.http.HttpServletRequest;
 import java.util.List;
-import java.util.Objects;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springdoc.core.annotations.ParameterObject;
-import org.springframework.beans.factory.annotation.Qualifier;
-import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -43,15 +32,8 @@ import org.springframework.web.bind.annotation.RestController;
 @RequiredArgsConstructor
 public class BoardController {
 
-    private static final String NON_VIEW_KEY = null;
-
-    @Qualifier("defaultRedisTemplate")
-    private final RedisTemplate<String, Object> redisTemplate;
     private final ResponseService responseService;
     private final BoardService boardService;
-    private final ProductService productService;
-    private final StoreService storeService;
-    private final ReviewService reviewService;
 
     @Operation(summary = "게시글 전체 조회")
     @ApiResponse(
@@ -107,71 +89,6 @@ public class BoardController {
         BoardCustomPage<List<BoardResponseDto>> boardResponseDto = boardService.getPostInFolder(
             memberId, sort, folderId, cursorId);
         return responseService.getSingleResult(boardResponseDto);
-    }
-
-    @Operation(summary = "스토어 조회")
-    @GetMapping("/{boardId}/store")
-    public CommonResult getStoreInfoInBoardDetail(
-        @PathVariable("boardId")
-        Long boardId,
-        @AuthenticationPrincipal
-        Long memberId
-    ) {
-        StoreDto storeDto = storeService.getStoreDtoByBoardId(memberId, boardId);
-        return responseService.getSingleResult(storeDto);
-    }
-
-    @Operation(summary = "게시판 조회")
-    @GetMapping("/{boardId}")
-    public CommonResult getBoardDetailResponse(
-        @PathVariable("boardId")
-        Long boardId,
-        @AuthenticationPrincipal
-        Long memberId,
-        HttpServletRequest httpServletRequest
-    ) {
-        String viewCountKey = getViewCountKey(boardId, httpServletRequest);
-        if (Boolean.TRUE.equals(redisTemplate.hasKey(viewCountKey))) {
-            BoardImageDetailResponse response = boardService.getBoardDtos(memberId, boardId,
-                NON_VIEW_KEY);
-            return responseService.getSingleResult(response);
-        }
-        BoardImageDetailResponse response = boardService.getBoardDtos(memberId, boardId,
-            viewCountKey);
-        return responseService.getSingleResult(response);
-    }
-
-    @Operation(summary = "상품 조회")
-    @GetMapping("/{boardId}/product")
-    public CommonResult getProductResponse(
-        @PathVariable("boardId")
-        Long boardId,
-        @AuthenticationPrincipal
-        Long memberId
-    ) {
-        if (Objects.nonNull(memberId)) {
-            ProductResponse productResponse = productService.getProductResponseWithPush(memberId,
-                boardId);
-            return responseService.getSingleResult(productResponse);
-        }
-
-        ProductResponse productResponse = productService.getProductResponse(boardId);
-        return responseService.getSingleResult(productResponse);
-    }
-
-    @Operation(summary = "리뷰 조회")
-    @GetMapping("/{boardId}/review")
-    public CommonResult getReviewResponse(
-        @PathVariable("boardId")
-        Long boardId) {
-        SummarizedReviewResponse response = reviewService.getSummarizedReview(boardId);
-
-        return responseService.getSingleResult(response);
-    }
-
-    private static String getViewCountKey(Long boardId, HttpServletRequest httpServletRequest) {
-        String ipAddress = httpServletRequest.getRemoteAddr();
-        return "VIEW:" + boardId + ":" + ipAddress;
     }
 
     private SortType settingDefaultSortTypeIfNull(SortType sort) {
