+++ src/main/java/com/bbangle/bbangle/board/controller/BoardDetailController.java
@@ -1,11 +1,21 @@
 package com.bbangle.bbangle.board.controller;
 
+import com.bbangle.bbangle.board.dto.BoardImageDetailResponse;
 import com.bbangle.bbangle.board.dto.SimilarityBoardResponse;
+import com.bbangle.bbangle.board.dto.orders.ProductResponse;
 import com.bbangle.bbangle.board.service.BoardDetailService;
+import com.bbangle.bbangle.board.service.ProductService;
 import com.bbangle.bbangle.common.dto.CommonResult;
 import com.bbangle.bbangle.common.service.ResponseService;
+import com.bbangle.bbangle.review.dto.SummarizedReviewResponse;
+import com.bbangle.bbangle.review.service.ReviewService;
+import com.bbangle.bbangle.store.dto.StoreDto;
+import com.bbangle.bbangle.store.service.StoreService;
+import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
+import jakarta.servlet.http.HttpServletRequest;
 import java.util.List;
+import java.util.Objects;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -19,8 +29,37 @@ import org.springframework.web.bind.annotation.RestController;
 @RequiredArgsConstructor
 public class BoardDetailController {
 
+    private final StoreService storeService;
     private final BoardDetailService boardDetailService;
     private final ResponseService responseService;
+    private final ProductService productService;
+    private final ReviewService reviewService;
+
+    @Operation(summary = "게시판 조회")
+    @GetMapping("/{boardId}")
+    public CommonResult getBoardDetailResponse(
+        @PathVariable("boardId")
+        Long boardId,
+        @AuthenticationPrincipal
+        Long memberId,
+        HttpServletRequest httpServletRequest
+    ) {
+        BoardImageDetailResponse response = boardDetailService.getBoardDtos(memberId, boardId,
+            httpServletRequest.getRemoteAddr());
+        return responseService.getSingleResult(response);
+    }
+
+    @Operation(summary = "스토어 조회")
+    @GetMapping("/{boardId}/store")
+    public CommonResult getStoreInfoInBoardDetail(
+        @PathVariable("boardId")
+        Long boardId,
+        @AuthenticationPrincipal
+        Long memberId
+    ) {
+        StoreDto storeDto = storeService.getStoreDtoByBoardId(memberId, boardId);
+        return responseService.getSingleResult(storeDto);
+    }
 
     @GetMapping("/{boardId}/similar_board")
     public CommonResult getCount(
@@ -32,6 +71,34 @@ public class BoardDetailController {
         List<SimilarityBoardResponse> similarityBoardResponses = boardDetailService.getSimilarityBoardResponses(
             memberId, boardId);
         return responseService.getSingleResult(similarityBoardResponses);
+    }
+
+    @Operation(summary = "상품 조회")
+    @GetMapping("/{boardId}/product")
+    public CommonResult getProductResponse(
+        @PathVariable("boardId")
+        Long boardId,
+        @AuthenticationPrincipal
+        Long memberId
+    ) {
+        if (Objects.nonNull(memberId)) {
+            ProductResponse productResponse = productService.getProductResponseWithPush(memberId,
+                boardId);
+            return responseService.getSingleResult(productResponse);
+        }
+
+        ProductResponse productResponse = productService.getProductResponse(boardId);
+        return responseService.getSingleResult(productResponse);
+    }
+
+    @Operation(summary = "리뷰 조회")
+    @GetMapping("/{boardId}/review")
+    public CommonResult getReviewResponse(
+        @PathVariable("boardId")
+        Long boardId) {
+        SummarizedReviewResponse response = reviewService.getSummarizedReview(boardId);
+
+        return responseService.getSingleResult(response);
     }
 
 }
